Call __open before calling open. Reset device once open.
__open seems to fix the case where open throws a LIBUSB_ERROR_NOT_FOUND
error. This has happened to me a couple of times after plugging the
device back in or when waking the computer from sleep. Probably related
to: https://github.com/tessel/node-usb/issues/61

I think I added the reset at some point when first trying to figure this
out. Doesn't seem to cause any problems, and it seems like a sensible
thing to do.